+++ CHANGELOG.md
@@ -4,6 +4,7 @@ SHORT FUTURE IMPROVEMENTS (post v1.0)
 - Sync and delete propagation function merge (master and slave functions are the same, reduces code size and maintain effort)
 - Tree function merge (current and after tree functions are the same except for output filename and logging, reduces code size and maintain effort)
 - Tree functions execute piped commands (grep, awk) on master when launched on remote slave which can cause more bandwith usage
+- Daemonize osync --on-changes mode
 
 FAR FUTURE IMPROVEMENTS
 -----------------------
@@ -20,6 +21,7 @@ KNOWN ISSUES
 RECENT CHANGES
 --------------
 
+- Added monitor mode, which will launch a sync task upon file operations on master replica
 - Changed conf file default format for ssh uri (old format is still compatible)
 - Added ssh uri support for slave replicas
 - Improved execution hooks logs